feat: move connection delete action to connection list
